requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.5.2
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Cython==0.29.28
cython-bbox==0.1.3
easydict==1.9
filterpy==1.4.5
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
google-auth==2.6.4
google-auth-oauthlib==0.4.6
grpcio==1.44.0
hydra-core==1.1.2
idna==3.3
imageio==2.16.2
imgaug==0.4.0
importlib-metadata==4.11.3
importlib-resources==5.2.3
joblib==1.1.0
kiwisolver==1.4.2
kornia==0.5.0
lap==0.4.0
loguru==0.6.0
Markdown==3.3.6
matplotlib==3.5.1
mmdet==2.23.0
multidict==6.0.2
networkx==2.8
ninja==1.10.2.3
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyDeprecate==0.3.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch-lightning==1.6.1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2022.4.8
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.26.9
webdataset==0.2.5
Werkzeug==2.1.1
wget==3.2
wldhx.yadisk-direct==0.0.6
yapf==0.32.0
yarl==1.7.2
zipp==3.8.0

torch~=1.13.1
tensorflow~=2.11.0